feat(ui): simplify and consolidate video capture logic
