+++ invokeai/frontend/web/src/features/video/components/VideoPlayer.tsx
@@ -1,10 +1,8 @@
 import { Flex } from '@invoke-ai/ui-library';
-import { $authToken } from 'app/store/nanostores/authToken';
 import { useVideoContextMenu } from 'features/gallery/components/ContextMenu/VideoContextMenu';
 import { useVideoViewerContext } from 'features/video/context/VideoViewerContext';
-import type { MediaStateOwner } from 'media-chrome/dist/media-store/state-mediator.js';
 import { MediaController } from 'media-chrome/react';
-import { memo, useCallback, useRef } from 'react';
+import { memo, useRef } from 'react';
 import ReactPlayer from 'react-player';
 import type { VideoDTO } from 'services/api/types';
 
@@ -16,17 +14,8 @@ interface VideoPlayerProps {
 
 export const VideoPlayer = memo(({ videoDTO }: VideoPlayerProps) => {
   const ref = useRef<HTMLDivElement>(null);
-  const reactPlayerRef = useRef<MediaStateOwner | null>(null);
   useVideoContextMenu(videoDTO, ref);
-  const { setVideoRef } = useVideoViewerContext();
-
-  const handleMediaRef = useCallback(
-    (mediaEl: MediaStateOwner | null | undefined) => {
-      reactPlayerRef.current = mediaEl || null;
-      setVideoRef(mediaEl || null);
-    },
-    [setVideoRef]
-  );
+  const { videoRef } = useVideoViewerContext();
 
   return (
     <Flex ref={ref} w="full" h="full" flexDirection="column" gap={4} alignItems="center" justifyContent="center">
@@ -38,13 +27,13 @@ export const VideoPlayer = memo(({ videoDTO }: VideoPlayerProps) => {
       >
         <ReactPlayer
           slot="media"
-          ref={handleMediaRef}
+          ref={videoRef}
           src={videoDTO.video_url}
           controls={false}
           width={videoDTO.width}
           height={videoDTO.height}
           pip={false}
-          crossOrigin={$authToken.get() ? 'use-credentials' : 'anonymous'}
+          // crossOrigin={$authToken.get() ? 'use-credentials' : 'anonymous'}
         />
 
         <VideoPlayerControls />
